requirements.txt
Flask==2.3.3
requests==2.31.0
pymysql==1.1.0
python-dotenv==1.0.1
APScheduler==3.10.4